rename languageServiceWrapper to languageServerManager
