+++ package/src/languageServiceManager/kustoMonarchLanguageDefinition.ts
