+++ package/src/languageServiceManager/schema.ts
