+++ package/src/languageServiceManager/settings.ts
